Only use Python tool names for valid resolves
This required moving the rule to resolve ResolvePexConfig to avoid a circular import
